refactor: API 하나의 파일로 통합 및 URL 상수화
- MyPage를 MyProfile로 수정
