+++ front-end/legeno-around-here/src/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 import Join from './components/pages/Join';
 import Login from './components/pages/Login';
 import Posting from './components/pages/Posting';
-import MyPage from './components/pages/MyPage';
+import MyProfile from './components/pages/MyProfile';
 import Sector from './components/pages/Sector';
 import PostDetail from './components/pages/PostDetail';
 import Home from './components/pages/Home';
@@ -25,7 +25,7 @@ function App() {
         <Route path="/" exact component={Home} />
         <Route path="/join" exact component={Join} />
         <Route path="/login" exact component={Login} />
-        <Route path="/mypage" exact component={MyPage} />
+        <Route path="/myProfile" exact component={MyProfile} />
         <Route path="/posting" exact component={Posting} />
         <Route path="/sector" exact component={Sector} />
         <Route path="/posts/:postId" exact component={PostDetail} />
