+++ front-end/legeno-around-here/src/components/myProfile/PrivacySection.js
